backend/requirements.txt
# ─── Django Core ─────────────────────────────────────────────────────────────
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
django-environ==0.11.2
django-redis==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
drf-spectacular==0.27.2  # OpenAPI docs

# ─── ASGI / Channels ─────────────────────────────────────────────────────────
channels==4.1.0
channels-redis==4.2.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0

# ─── Database ────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9
django-extensions==3.2.3
# PostGIS
# GDAL is installed system-level via apt

# ─── Celery ──────────────────────────────────────────────────────────────────
celery[redis]==5.4.0
kombu==5.3.5
flower==2.0.1

# ─── Geo / Map ───────────────────────────────────────────────────────────────
h3==3.7.7
shapely==2.0.4
pyproj==3.6.1
geojson==3.1.0

# ─── Blockchain / Web3 ───────────────────────────────────────────────────────
web3==6.19.0
eth-account==0.12.3
eth-typing==4.1.0

# ─── MongoDB ─────────────────────────────────────────────────────────────────
pymongo==4.7.2
motor==3.4.0

# ─── Kafka ───────────────────────────────────────────────────────────────────
kafka-python==2.0.2

# ─── Security ────────────────────────────────────────────────────────────────
django-ratelimit==4.1.0
cryptography==42.0.7
PyJWT==2.8.0

# ─── Utils ───────────────────────────────────────────────────────────────────
Pillow==10.3.0
requests==2.32.2
httpx==0.27.0
pydantic==2.7.1
orjson==3.10.3
python-dateutil==2.9.0
numpy==1.26.4
sentry-sdk[django]==2.3.1
django-storages[s3]==1.14.3
boto3==1.34.110
